backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
boto3>=1.34.0
neo4j>=5.19.0
scikit-learn>=1.4.0
xgboost>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
python-multipart>=0.0.9
numpy>=1.26.0
